Move layouts and block-renderer from components to src folder
