+++ src/block-renderer/BlockRenderer.template.ts
@@ -1,10 +1,10 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import type { TemplateMap } from '@muban/template';
 import { html } from '@muban/template';
-import { renderLazyComponentTemplate } from '../../utils/createComponentRenderer';
-import { imageTestTemplate } from '../image-test/ImageTest.template';
-import { toggleExpandTemplate } from '../atoms/toggle-expand/ToggleExpand.template';
-import { videoTestTemplate } from '../video-test/VideoTest.template';
+import { renderLazyComponentTemplate } from '../utils/createComponentRenderer';
+import { imageTestTemplate } from '../components/image-test/ImageTest.template';
+import { toggleExpandTemplate } from '../components/atoms/toggle-expand/ToggleExpand.template';
+import { videoTestTemplate } from '../components/video-test/VideoTest.template';
 
 const componentMap = {
   'toggle-expand': toggleExpandTemplate,
