+++ src/block-renderer/BlockRenderer.ts
@@ -1,13 +1,13 @@
 import { defineComponent, lazy } from '@muban/muban';
-import { ImageTest } from '../image-test/ImageTest';
-import { VideoTest } from '../video-test/VideoTest';
+import { ImageTest } from '../components/image-test/ImageTest';
+import { VideoTest } from '../components/video-test/VideoTest';
 
 export const BlockRenderer = defineComponent({
   name: 'block-renderer',
   components: [
     lazy(
       'toggle-expand',
-      () => import(/* webpackExports: "lazy" */ '../atoms/toggle-expand/ToggleExpand'),
+      () => import(/* webpackExports: "lazy" */ '../components/atoms/toggle-expand/ToggleExpand'),
     ),
     ImageTest,
     VideoTest,
